refactor: simplify build and packaging of list-group components
