+++ packages/components/list-group/bs-list-group-action.js
@@ -1,7 +1,7 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupActionCss } from './bs-list-group-action.css.js';
-import { BsListGroupCommonCss } from './bs-list-group-common.css.js';
+import { BsListGroupActionCss } from './css/bs-list-group-action.css.js';
+import { BsListGroupCommonCss } from './css/bs-list-group-common.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsListGroupAction extends LitElement {
