+++ packages/components/list-group/bs-list-group-badge-item.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupItemCommonCss } from './bs-list-group-item-common.css.js';
+import { BsListGroupItemCommonCss } from './css/bs-list-group-item-common.css.js';
 import { BsFlexJustifyCss } from '@lit-element-bootstrap/utilities';
 import { BsFlexDisplayCss } from '@lit-element-bootstrap/utilities';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
