+++ packages/components/list-group/bs-list-group-flush.js
@@ -1,16 +1,16 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupCss } from './bs-list-group.css.js';
-import { BsListGroupCommonCss } from './bs-list-group-common.css.js';
+import { BsListGroupFlushCss } from './css/bs-list-group-flush.css.js';
+import { BsListGroupCommonCss } from './css/bs-list-group-common.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
-export class BsListGroup extends LitElement {
+export class BsListGroupFlush extends LitElement {
 
     static get styles() {
         return [
             BsContentRebootCss,
             BsListGroupCommonCss,
-            BsListGroupCss
+            BsListGroupFlushCss
         ];
     }
 
@@ -23,5 +23,5 @@ export class BsListGroup extends LitElement {
     }
 };
 
-if(!window.customElements.get('bs-list-group'))
-    window.customElements.define('bs-list-group', BsListGroup);
+if(!window.customElements.get('bs-list-group-flush'))
+    window.customElements.define('bs-list-group-flush', BsListGroupFlush);
