+++ packages/components/list-group/bs-list-group-item-action-button.js
@@ -1,7 +1,7 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupItemCommonCss } from './bs-list-group-item-common.css.js';
-import { BsListGroupItemActionCommonCss } from './bs-list-group-item-action-common.css.js';
+import { BsListGroupItemCommonCss } from './css/bs-list-group-item-common.css.js';
+import { BsListGroupItemActionCommonCss } from './css/bs-list-group-item-action-common.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 import { BsContentTypographyCss } from '@lit-element-bootstrap/content';
