+++ packages/components/list-group/bs-list-group-item.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupItemCommonCss } from './bs-list-group-item-common.css.js';
+import { BsListGroupItemCommonCss } from './css/bs-list-group-item-common.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsListGroupItem extends LitElement {
