+++ packages/components/list-group/bs-list-group.js
@@ -1,16 +1,16 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsListGroupFlushCss } from './bs-list-group-flush.css.js';
-import { BsListGroupCommonCss } from './bs-list-group-common.css.js';
+import { BsListGroupCss } from './css/bs-list-group.css.js';
+import { BsListGroupCommonCss } from './css/bs-list-group-common.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
-export class BsListGroupFlush extends LitElement {
+export class BsListGroup extends LitElement {
 
     static get styles() {
         return [
             BsContentRebootCss,
             BsListGroupCommonCss,
-            BsListGroupFlushCss
+            BsListGroupCss
         ];
     }
 
@@ -23,5 +23,5 @@ export class BsListGroupFlush extends LitElement {
     }
 };
 
-if(!window.customElements.get('bs-list-group-flush'))
-    window.customElements.define('bs-list-group-flush', BsListGroupFlush);
+if(!window.customElements.get('bs-list-group'))
+    window.customElements.define('bs-list-group', BsListGroup);
